a simple function to reverse a given number
This idea for this approach is to repeatedly extract the last digit of num using the modulus operator (n % 10) and appending it to the reverse number (reverseDigit). After extracting the digit, the number num is reduced by dividing it by 10 (num = num / 10). This process continues until n becomes 0. Finally, the reversed number (reveseDigit) is returned.