simple function to reverse digits of a number
This idea for this approach is to repeatedly extract the last digit of num using the modulus operator (num % 10) and appending it to the reverse number (rev). After extracting the digit, the number num is reduced by dividing it by 10 (num = num / 10). This process continues until num becomes 0. Finally, the reversed number (rev) is returned.